create admin-data helper functions
