+++ public/sw.js
@@ -1,1 +1,1 @@
-if(!self.define){let e,s={};const c=(c,t)=>(c=new URL(c+".js",t).href,s[c]||new Promise((s=>{if("document"in self){const e=document.createElement("script");e.src=c,e.onload=s,document.head.appendChild(e)}else e=c,importScripts(c),s()})).then((()=>{let e=s[c];if(!e)throw new Error(`Module ${c} didn’t register its module`);return e})));self.define=(t,a)=>{const n=e||("document"in self?document.currentScript.src:"")||location.href;if(s[n])return;let i={};const r=e=>c(e,n),l={module:{uri:n},exports:i,require:r};s[n]=Promise.all(t.map((e=>l[e]||r(e)))).then((e=>(a(...e),i)))}}define(["./workbox-1bb06f5e"],(function(e){"use strict";importScripts(),self.skipWaiting(),e.clientsClaim(),e.precacheAndRoute([{url:"/_next/app-build-manifest.json",revision:"dbae2bccff4e61e6d9d75882485d51ba"},{url:"/_next/static/chunks/1-6188ef3660cdce7f.js",revision:"vgzcZvR1UHtxLLQIlPWN8"},{url:"/_next/static/chunks/126.12885d22201eec05.js",revision:"12885d22201eec05"},{url:"/_next/static/chunks/160-e1e398bb3fd4f27b.js",revision:"vgzcZvR1UHtxLLQIlPWN8"},{url:"/_next/static/chunks/208-ccac7bd797189028.js",revision:"vgzcZvR1UHtxLLQIlPWN8"},{url:"/_next/static/chunks/341.df3329d77a5faa19.js",revision:"df3329d77a5faa19"},{url:"/_next/static/chunks/352-ca7b0e1cd96c0ec6.js",revision:"vgzcZvR1UHtxLLQIlPWN8"},{url:"/_next/static/chunks/45-75bf477b4aec45a4.js",revision:"vgzcZvR1UHtxLLQIlPWN8"},{url:"/_next/static/chunks/453-f61403d3cfd4e732.js",revision:"vgzcZvR1UHtxLLQIlPWN8"},{url:"/_next/static/chunks/472.a3826d29d6854395.js",revision:"a3826d29d6854395"},{url:"/_next/static/chunks/4bd1b696-482475e2590df3e4.js",revision:"vgzcZvR1UHtxLLQIlPWN8"},{url:"/_next/static/chunks/63-b2edba4ece5b8df3.js",revision:"vgzcZvR1UHtxLLQIlPWN8"},{url:"/_next/static/chunks/656-5715616fea404d5a.js",revision:"vgzcZvR1UHtxLLQIlPWN8"},{url:"/_next/static/chunks/684-cb01d885a4e91b9a.js",revision:"vgzcZvR1UHtxLLQIlPWN8"},{url:"/_next/static/chunks/689-5763385fbfae106d.js",revision:"vgzcZvR1UHtxLLQIlPWN8"},{url:"/_next/static/chunks/696-b77bf4933ae34647.js",revision:"vgzcZvR1UHtxLLQIlPWN8"},{url:"/_next/static/chunks/740-150a7866895b84f7.js",revision:"vgzcZvR1UHtxLLQIlPWN8"},{url:"/_next/static/chunks/741-caa422bfe08fcf0f.js",revision:"vgzcZvR1UHtxLLQIlPWN8"},{url:"/_next/static/chunks/935-31d4909e3a8a89e9.js",revision:"vgzcZvR1UHtxLLQIlPWN8"},{url:"/_next/static/chunks/959-ed708a2a4c069ec3.js",revision:"vgzcZvR1UHtxLLQIlPWN8"},{url:"/_next/static/chunks/app/%5Blocale%5D/about/page-bf54011b93f9b167.js",revision:"vgzcZvR1UHtxLLQIlPWN8"},{url:"/_next/static/chunks/app/%5Blocale%5D/admin/messages/page-f9872960381e96bc.js",revision:"vgzcZvR1UHtxLLQIlPWN8"},{url:"/_next/static/chunks/app/%5Blocale%5D/admin/page-b7fb3f3a7d4dc75d.js",revision:"vgzcZvR1UHtxLLQIlPWN8"},{url:"/_next/static/chunks/app/%5Blocale%5D/contact/page-b91b353e5435104a.js",revision:"vgzcZvR1UHtxLLQIlPWN8"},{url:"/_next/static/chunks/app/%5Blocale%5D/layout-a32b5453cad7cec6.js",revision:"vgzcZvR1UHtxLLQIlPWN8"},{url:"/_next/static/chunks/app/%5Blocale%5D/login/page-4be9d4044ce103f6.js",revision:"vgzcZvR1UHtxLLQIlPWN8"},{url:"/_next/static/chunks/app/%5Blocale%5D/my-bookings/page-7fe9946f97979b65.js",revision:"vgzcZvR1UHtxLLQIlPWN8"},{url:"/_next/static/chunks/app/%5Blocale%5D/page-7530661229dddf74.js",revision:"vgzcZvR1UHtxLLQIlPWN8"},{url:"/_next/static/chunks/app/%5Blocale%5D/reset-password/page-30cd9e1809f626a7.js",revision:"vgzcZvR1UHtxLLQIlPWN8"},{url:"/_next/static/chunks/app/%5Blocale%5D/schedule/page-3883117b734df8f1.js",revision:"vgzcZvR1UHtxLLQIlPWN8"},{url:"/_next/static/chunks/app/%5Blocale%5D/signup/page-08010f7328324367.js",revision:"vgzcZvR1UHtxLLQIlPWN8"},{url:"/_next/static/chunks/app/%5Blocale%5D/update-password/page-382bce8eb67e8959.js",revision:"vgzcZvR1UHtxLLQIlPWN8"},{url:"/_next/static/chunks/app/_not-found/page-8bf4098514b7a9d9.js",revision:"vgzcZvR1UHtxLLQIlPWN8"},{url:"/_next/static/chunks/app/api/auth/callback/route-201f069797d71123.js",revision:"vgzcZvR1UHtxLLQIlPWN8"},{url:"/_next/static/chunks/app/api/auth/login/route-daea7b7cc7977f7b.js",revision:"vgzcZvR1UHtxLLQIlPWN8"},{url:"/_next/static/chunks/app/api/auth/logout/route-1b8206640917b294.js",revision:"vgzcZvR1UHtxLLQIlPWN8"},{url:"/_next/static/chunks/app/api/auth/reset-password/route-112e805045b7061c.js",revision:"vgzcZvR1UHtxLLQIlPWN8"},{url:"/_next/static/chunks/app/api/auth/signup/route-d2d953634e8d1154.js",revision:"vgzcZvR1UHtxLLQIlPWN8"},{url:"/_next/static/chunks/app/api/profile/route-565d0c93eaac802e.js",revision:"vgzcZvR1UHtxLLQIlPWN8"},{url:"/_next/static/chunks/app/classes/page-7d185c3790fe32d5.js",revision:"vgzcZvR1UHtxLLQIlPWN8"},{url:"/_next/static/chunks/app/error/page-c6e32d8412b75e55.js",revision:"vgzcZvR1UHtxLLQIlPWN8"},{url:"/_next/static/chunks/app/layout-cad14c5e5c8c4331.js",revision:"vgzcZvR1UHtxLLQIlPWN8"},{url:"/_next/static/chunks/framework-f593a28cde54158e.js",revision:"vgzcZvR1UHtxLLQIlPWN8"},{url:"/_next/static/chunks/main-8823abe1b59a826e.js",revision:"vgzcZvR1UHtxLLQIlPWN8"},{url:"/_next/static/chunks/main-app-8af56825497c9159.js",revision:"vgzcZvR1UHtxLLQIlPWN8"},{url:"/_next/static/chunks/pages/_app-92f2aae776f86b9c.js",revision:"vgzcZvR1UHtxLLQIlPWN8"},{url:"/_next/static/chunks/pages/_error-71d2b6a7b832d02a.js",revision:"vgzcZvR1UHtxLLQIlPWN8"},{url:"/_next/static/chunks/polyfills-42372ed130431b0a.js",revision:"846118c33b2c0e922d7b3a7676f81f6f"},{url:"/_next/static/chunks/webpack-df1d7c9c702436a3.js",revision:"vgzcZvR1UHtxLLQIlPWN8"},{url:"/_next/static/css/4405f6908fc2a9d0.css",revision:"4405f6908fc2a9d0"},{url:"/_next/static/media/569ce4b8f30dc480-s.p.woff2",revision:"ef6cefb32024deac234e82f932a95cbd"},{url:"/_next/static/media/747892c23ea88013-s.woff2",revision:"a0761690ccf4441ace5cec893b82d4ab"},{url:"/_next/static/media/8d697b304b401681-s.woff2",revision:"cc728f6c0adb04da0dfcb0fc436a8ae5"},{url:"/_next/static/media/93f479601ee12b01-s.p.woff2",revision:"da83d5f06d825c5ae65b7cca706cb312"},{url:"/_next/static/media/9610d9e46709d722-s.woff2",revision:"7b7c0ef93df188a852344fc272fc096b"},{url:"/_next/static/media/ba015fad6dcf6784-s.woff2",revision:"8ea4f719af3312a055caf09f34c89a77"},{url:"/_next/static/vgzcZvR1UHtxLLQIlPWN8/_buildManifest.js",revision:"5c990e60adf8bbfed70ad21d63652322"},{url:"/_next/static/vgzcZvR1UHtxLLQIlPWN8/_ssgManifest.js",revision:"b6652df95db52feb4daf4eca35380933"},{url:"/aleksandra_pavic.webp",revision:"f707ea48a258a2cadc3e42c3601cac80"},{url:"/aleksandra_sikirica.webp",revision:"f707ea48a258a2cadc3e42c3601cac80"},{url:"/background_studio.webp",revision:"19f825198f53f1a001c1842a3342c275"},{url:"/biljana_zivkovic.webp",revision:"578194abcbf07bd327f4641285f51d12"},{url:"/bojana_milic.webp",revision:"f707ea48a258a2cadc3e42c3601cac80"},{url:"/equipment.webp",revision:"3cbb6399131593a0d490a00f5cce218e"},{url:"/facility.webp",revision:"bd41899f749c2603e55b052e314275d3"},{url:"/icon-192.png",revision:"b72f5895c36ef1697d2e4018656d5e78"},{url:"/icon-512.png.png",revision:"b72f5895c36ef1697d2e4018656d5e78"},{url:"/manifest.json",revision:"dcf4fa0d8bd1f5715b7fbf01f139538b"}],{ignoreURLParametersMatching:[]}),e.cleanupOutdatedCaches(),e.registerRoute("/",new e.NetworkFirst({cacheName:"start-url",plugins:[{cacheWillUpdate:async({request:e,response:s,event:c,state:t})=>s&&"opaqueredirect"===s.type?new Response(s.body,{status:200,statusText:"OK",headers:s.headers}):s}]}),"GET"),e.registerRoute(/^https:\/\/fonts\.(?:gstatic)\.com\/.*/i,new e.CacheFirst({cacheName:"google-fonts-webfonts",plugins:[new e.ExpirationPlugin({maxEntries:4,maxAgeSeconds:31536e3})]}),"GET"),e.registerRoute(/^https:\/\/fonts\.(?:googleapis)\.com\/.*/i,new e.StaleWhileRevalidate({cacheName:"google-fonts-stylesheets",plugins:[new e.ExpirationPlugin({maxEntries:4,maxAgeSeconds:604800})]}),"GET"),e.registerRoute(/\.(?:eot|otf|ttc|ttf|woff|woff2|font.css)$/i,new e.StaleWhileRevalidate({cacheName:"static-font-assets",plugins:[new e.ExpirationPlugin({maxEntries:4,maxAgeSeconds:604800})]}),"GET"),e.registerRoute(/\.(?:jpg|jpeg|gif|png|svg|ico|webp)$/i,new e.StaleWhileRevalidate({cacheName:"static-image-assets",plugins:[new e.ExpirationPlugin({maxEntries:64,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\/_next\/image\?url=.+$/i,new e.StaleWhileRevalidate({cacheName:"next-image",plugins:[new e.ExpirationPlugin({maxEntries:64,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\.(?:mp3|wav|ogg)$/i,new e.CacheFirst({cacheName:"static-audio-assets",plugins:[new e.RangeRequestsPlugin,new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\.(?:mp4)$/i,new e.CacheFirst({cacheName:"static-video-assets",plugins:[new e.RangeRequestsPlugin,new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\.(?:js)$/i,new e.StaleWhileRevalidate({cacheName:"static-js-assets",plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\.(?:css|less)$/i,new e.StaleWhileRevalidate({cacheName:"static-style-assets",plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\/_next\/data\/.+\/.+\.json$/i,new e.StaleWhileRevalidate({cacheName:"next-data",plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\.(?:json|xml|csv)$/i,new e.NetworkFirst({cacheName:"static-data-assets",plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute((({url:e})=>{if(!(self.origin===e.origin))return!1;const s=e.pathname;return!s.startsWith("/api/auth/")&&!!s.startsWith("/api/")}),new e.NetworkFirst({cacheName:"apis",networkTimeoutSeconds:10,plugins:[new e.ExpirationPlugin({maxEntries:16,maxAgeSeconds:86400})]}),"GET"),e.registerRoute((({url:e})=>{if(!(self.origin===e.origin))return!1;return!e.pathname.startsWith("/api/")}),new e.NetworkFirst({cacheName:"others",networkTimeoutSeconds:10,plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute((({url:e})=>!(self.origin===e.origin)),new e.NetworkFirst({cacheName:"cross-origin",networkTimeoutSeconds:10,plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:3600})]}),"GET")}));
+if(!self.define){let e,s={};const a=(a,n)=>(a=new URL(a+".js",n).href,s[a]||new Promise((s=>{if("document"in self){const e=document.createElement("script");e.src=a,e.onload=s,document.head.appendChild(e)}else e=a,importScripts(a),s()})).then((()=>{let e=s[a];if(!e)throw new Error(`Module ${a} didn’t register its module`);return e})));self.define=(n,i)=>{const c=e||("document"in self?document.currentScript.src:"")||location.href;if(s[c])return;let t={};const r=e=>a(e,c),b={module:{uri:c},exports:t,require:r};s[c]=Promise.all(n.map((e=>b[e]||r(e)))).then((e=>(i(...e),t)))}}define(["./workbox-1bb06f5e"],(function(e){"use strict";importScripts(),self.skipWaiting(),e.clientsClaim(),e.precacheAndRoute([{url:"/_next/app-build-manifest.json",revision:"490e2f0e26701b39738b4ccd37534df2"},{url:"/_next/static/chunks/126.12885d22201eec05.js",revision:"12885d22201eec05"},{url:"/_next/static/chunks/157-d2f2c83b4280210a.js",revision:"fWLHbWa9rYvShn0sm2bN9"},{url:"/_next/static/chunks/222-b2e243f913601c6a.js",revision:"fWLHbWa9rYvShn0sm2bN9"},{url:"/_next/static/chunks/341.df3329d77a5faa19.js",revision:"df3329d77a5faa19"},{url:"/_next/static/chunks/352-ca7b0e1cd96c0ec6.js",revision:"fWLHbWa9rYvShn0sm2bN9"},{url:"/_next/static/chunks/372-8f3151280d40b3d3.js",revision:"fWLHbWa9rYvShn0sm2bN9"},{url:"/_next/static/chunks/410-ccedb0699d5825f9.js",revision:"fWLHbWa9rYvShn0sm2bN9"},{url:"/_next/static/chunks/453-f61403d3cfd4e732.js",revision:"fWLHbWa9rYvShn0sm2bN9"},{url:"/_next/static/chunks/472.a3826d29d6854395.js",revision:"a3826d29d6854395"},{url:"/_next/static/chunks/4bd1b696-482475e2590df3e4.js",revision:"fWLHbWa9rYvShn0sm2bN9"},{url:"/_next/static/chunks/63-f6b55a449939d09f.js",revision:"fWLHbWa9rYvShn0sm2bN9"},{url:"/_next/static/chunks/684-cb01d885a4e91b9a.js",revision:"fWLHbWa9rYvShn0sm2bN9"},{url:"/_next/static/chunks/689-5763385fbfae106d.js",revision:"fWLHbWa9rYvShn0sm2bN9"},{url:"/_next/static/chunks/740-150a7866895b84f7.js",revision:"fWLHbWa9rYvShn0sm2bN9"},{url:"/_next/static/chunks/753-2ac0a796f619b622.js",revision:"fWLHbWa9rYvShn0sm2bN9"},{url:"/_next/static/chunks/874-023775ccaca48b0c.js",revision:"fWLHbWa9rYvShn0sm2bN9"},{url:"/_next/static/chunks/935-31d4909e3a8a89e9.js",revision:"fWLHbWa9rYvShn0sm2bN9"},{url:"/_next/static/chunks/959-6f5ef5c38bee3a99.js",revision:"fWLHbWa9rYvShn0sm2bN9"},{url:"/_next/static/chunks/app/%5Blocale%5D/about/page-5e269358dc755a19.js",revision:"fWLHbWa9rYvShn0sm2bN9"},{url:"/_next/static/chunks/app/%5Blocale%5D/admin/messages/page-76d257ce907aa156.js",revision:"fWLHbWa9rYvShn0sm2bN9"},{url:"/_next/static/chunks/app/%5Blocale%5D/admin/page-cc4e40a6da6e51f7.js",revision:"fWLHbWa9rYvShn0sm2bN9"},{url:"/_next/static/chunks/app/%5Blocale%5D/contact/page-ca82dbfe3c79040c.js",revision:"fWLHbWa9rYvShn0sm2bN9"},{url:"/_next/static/chunks/app/%5Blocale%5D/layout-43c852241dd3e619.js",revision:"fWLHbWa9rYvShn0sm2bN9"},{url:"/_next/static/chunks/app/%5Blocale%5D/login/page-63b755d84e5d57f7.js",revision:"fWLHbWa9rYvShn0sm2bN9"},{url:"/_next/static/chunks/app/%5Blocale%5D/my-bookings/page-598e821a103b62de.js",revision:"fWLHbWa9rYvShn0sm2bN9"},{url:"/_next/static/chunks/app/%5Blocale%5D/page-d0d1544dc2a470db.js",revision:"fWLHbWa9rYvShn0sm2bN9"},{url:"/_next/static/chunks/app/%5Blocale%5D/price-list/page-d883e5ea97623c80.js",revision:"fWLHbWa9rYvShn0sm2bN9"},{url:"/_next/static/chunks/app/%5Blocale%5D/reset-password/page-91f68c01f5d449b9.js",revision:"fWLHbWa9rYvShn0sm2bN9"},{url:"/_next/static/chunks/app/%5Blocale%5D/schedule/page-b9885b461a1e95aa.js",revision:"fWLHbWa9rYvShn0sm2bN9"},{url:"/_next/static/chunks/app/%5Blocale%5D/signup/page-1cf0ef1bda957b36.js",revision:"fWLHbWa9rYvShn0sm2bN9"},{url:"/_next/static/chunks/app/%5Blocale%5D/update-password/page-02c1f36fb8c7defb.js",revision:"fWLHbWa9rYvShn0sm2bN9"},{url:"/_next/static/chunks/app/_not-found/page-8bf4098514b7a9d9.js",revision:"fWLHbWa9rYvShn0sm2bN9"},{url:"/_next/static/chunks/app/api/auth/callback/route-201f069797d71123.js",revision:"fWLHbWa9rYvShn0sm2bN9"},{url:"/_next/static/chunks/app/api/auth/login/route-daea7b7cc7977f7b.js",revision:"fWLHbWa9rYvShn0sm2bN9"},{url:"/_next/static/chunks/app/api/auth/logout/route-1b8206640917b294.js",revision:"fWLHbWa9rYvShn0sm2bN9"},{url:"/_next/static/chunks/app/api/auth/reset-password/route-112e805045b7061c.js",revision:"fWLHbWa9rYvShn0sm2bN9"},{url:"/_next/static/chunks/app/api/auth/signup/route-d2d953634e8d1154.js",revision:"fWLHbWa9rYvShn0sm2bN9"},{url:"/_next/static/chunks/app/api/profile/route-565d0c93eaac802e.js",revision:"fWLHbWa9rYvShn0sm2bN9"},{url:"/_next/static/chunks/app/classes/page-8c9668ec702eeac9.js",revision:"fWLHbWa9rYvShn0sm2bN9"},{url:"/_next/static/chunks/app/error/page-f13118d9a5b05d56.js",revision:"fWLHbWa9rYvShn0sm2bN9"},{url:"/_next/static/chunks/app/layout-0695456d595d91fc.js",revision:"fWLHbWa9rYvShn0sm2bN9"},{url:"/_next/static/chunks/framework-f593a28cde54158e.js",revision:"fWLHbWa9rYvShn0sm2bN9"},{url:"/_next/static/chunks/main-8823abe1b59a826e.js",revision:"fWLHbWa9rYvShn0sm2bN9"},{url:"/_next/static/chunks/main-app-8171f8b27e8fd6d3.js",revision:"fWLHbWa9rYvShn0sm2bN9"},{url:"/_next/static/chunks/pages/_app-92f2aae776f86b9c.js",revision:"fWLHbWa9rYvShn0sm2bN9"},{url:"/_next/static/chunks/pages/_error-71d2b6a7b832d02a.js",revision:"fWLHbWa9rYvShn0sm2bN9"},{url:"/_next/static/chunks/polyfills-42372ed130431b0a.js",revision:"846118c33b2c0e922d7b3a7676f81f6f"},{url:"/_next/static/chunks/webpack-df1d7c9c702436a3.js",revision:"fWLHbWa9rYvShn0sm2bN9"},{url:"/_next/static/css/6c2ae94d2dbae237.css",revision:"6c2ae94d2dbae237"},{url:"/_next/static/fWLHbWa9rYvShn0sm2bN9/_buildManifest.js",revision:"5c990e60adf8bbfed70ad21d63652322"},{url:"/_next/static/fWLHbWa9rYvShn0sm2bN9/_ssgManifest.js",revision:"b6652df95db52feb4daf4eca35380933"},{url:"/_next/static/media/569ce4b8f30dc480-s.p.woff2",revision:"ef6cefb32024deac234e82f932a95cbd"},{url:"/_next/static/media/747892c23ea88013-s.woff2",revision:"a0761690ccf4441ace5cec893b82d4ab"},{url:"/_next/static/media/8d697b304b401681-s.woff2",revision:"cc728f6c0adb04da0dfcb0fc436a8ae5"},{url:"/_next/static/media/93f479601ee12b01-s.p.woff2",revision:"da83d5f06d825c5ae65b7cca706cb312"},{url:"/_next/static/media/9610d9e46709d722-s.woff2",revision:"7b7c0ef93df188a852344fc272fc096b"},{url:"/_next/static/media/ba015fad6dcf6784-s.woff2",revision:"8ea4f719af3312a055caf09f34c89a77"},{url:"/aleksandra_pavic.webp",revision:"f707ea48a258a2cadc3e42c3601cac80"},{url:"/aleksandra_sikirica.webp",revision:"92dcb697eb1f23e740a66bcfe3d95ab3"},{url:"/background_studio.webp",revision:"19f825198f53f1a001c1842a3342c275"},{url:"/biljana_zivkovic.webp",revision:"578194abcbf07bd327f4641285f51d12"},{url:"/bojana_milic.webp",revision:"f707ea48a258a2cadc3e42c3601cac80"},{url:"/carousel1.webp",revision:"45fbfedcbf0e206da101f537636add66"},{url:"/carousel2.webp",revision:"bcd902b5dbbe72a3ffad9c9571577259"},{url:"/carousel3.webp",revision:"dda2636d6f5612e26902235d4aeb0e2e"},{url:"/equipment.webp",revision:"3cbb6399131593a0d490a00f5cce218e"},{url:"/facility.webp",revision:"bd41899f749c2603e55b052e314275d3"},{url:"/icon-192.png",revision:"b72f5895c36ef1697d2e4018656d5e78"},{url:"/icon-512.png.png",revision:"b72f5895c36ef1697d2e4018656d5e78"},{url:"/manifest.json",revision:"dcf4fa0d8bd1f5715b7fbf01f139538b"}],{ignoreURLParametersMatching:[]}),e.cleanupOutdatedCaches(),e.registerRoute("/",new e.NetworkFirst({cacheName:"start-url",plugins:[{cacheWillUpdate:async({request:e,response:s,event:a,state:n})=>s&&"opaqueredirect"===s.type?new Response(s.body,{status:200,statusText:"OK",headers:s.headers}):s}]}),"GET"),e.registerRoute(/^https:\/\/fonts\.(?:gstatic)\.com\/.*/i,new e.CacheFirst({cacheName:"google-fonts-webfonts",plugins:[new e.ExpirationPlugin({maxEntries:4,maxAgeSeconds:31536e3})]}),"GET"),e.registerRoute(/^https:\/\/fonts\.(?:googleapis)\.com\/.*/i,new e.StaleWhileRevalidate({cacheName:"google-fonts-stylesheets",plugins:[new e.ExpirationPlugin({maxEntries:4,maxAgeSeconds:604800})]}),"GET"),e.registerRoute(/\.(?:eot|otf|ttc|ttf|woff|woff2|font.css)$/i,new e.StaleWhileRevalidate({cacheName:"static-font-assets",plugins:[new e.ExpirationPlugin({maxEntries:4,maxAgeSeconds:604800})]}),"GET"),e.registerRoute(/\.(?:jpg|jpeg|gif|png|svg|ico|webp)$/i,new e.StaleWhileRevalidate({cacheName:"static-image-assets",plugins:[new e.ExpirationPlugin({maxEntries:64,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\/_next\/image\?url=.+$/i,new e.StaleWhileRevalidate({cacheName:"next-image",plugins:[new e.ExpirationPlugin({maxEntries:64,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\.(?:mp3|wav|ogg)$/i,new e.CacheFirst({cacheName:"static-audio-assets",plugins:[new e.RangeRequestsPlugin,new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\.(?:mp4)$/i,new e.CacheFirst({cacheName:"static-video-assets",plugins:[new e.RangeRequestsPlugin,new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\.(?:js)$/i,new e.StaleWhileRevalidate({cacheName:"static-js-assets",plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\.(?:css|less)$/i,new e.StaleWhileRevalidate({cacheName:"static-style-assets",plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\/_next\/data\/.+\/.+\.json$/i,new e.StaleWhileRevalidate({cacheName:"next-data",plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\.(?:json|xml|csv)$/i,new e.NetworkFirst({cacheName:"static-data-assets",plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute((({url:e})=>{if(!(self.origin===e.origin))return!1;const s=e.pathname;return!s.startsWith("/api/auth/")&&!!s.startsWith("/api/")}),new e.NetworkFirst({cacheName:"apis",networkTimeoutSeconds:10,plugins:[new e.ExpirationPlugin({maxEntries:16,maxAgeSeconds:86400})]}),"GET"),e.registerRoute((({url:e})=>{if(!(self.origin===e.origin))return!1;return!e.pathname.startsWith("/api/")}),new e.NetworkFirst({cacheName:"others",networkTimeoutSeconds:10,plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute((({url:e})=>!(self.origin===e.origin)),new e.NetworkFirst({cacheName:"cross-origin",networkTimeoutSeconds:10,plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:3600})]}),"GET")}));
